refactored code into modules
